geoloc dans navbar
